Fixed per repo standards including removing global vars, updating header and adding --scheme command line option.  Fixed filename.
